# src/main/java/io/swagger/api/SickApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import ru.fds.hrdepartment.dto.SickLeaveDto;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-09-11T10:14:57.918+03:00")

@Api(value = "sick", description = "the sick API")
public interface SickApi {

    @ApiOperation(value = "get sick leave by id", nickname = "getSickLeave", notes = "", response = SickLeaveDto.class, tags={ "sickLeaveController", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = SickLeaveDto.class),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/sick/{sickId}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<SickLeaveDto> getSickLeave(@ApiParam(value = "sickId",required=true) @PathVariable("sickId") Long sickId);


    @ApiOperation(value = "insert new sick leave", nickname = "insertSickLeave", notes = "", response = SickLeaveDto.class, tags={ "sickLeaveController", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = SickLeaveDto.class),
        @ApiResponse(code = 201, message = "Created"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/sick/new",
        produces = { "*/*" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<SickLeaveDto> insertSickLeave(@ApiParam(value = "sickLeaveDto" ,required=true )  @Valid @RequestBody SickLeaveDto sickLeaveDto);


    @ApiOperation(value = "update sick leave", nickname = "updateSickLeave", notes = "", response = SickLeaveDto.class, tags={ "sickLeaveController", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = SickLeaveDto.class),
        @ApiResponse(code = 201, message = "Created"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/sick/update",
        produces = { "*/*" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<SickLeaveDto> updateSickLeave(@ApiParam(value = "sickLeaveDto" ,required=true )  @Valid @RequestBody SickLeaveDto sickLeaveDto);

}