# src/main/resources/api.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
...
paths:
...
  /position/new:
    post:
      tags:
        - positionController
      summary: insert new position
      operationId: insertPosition
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: positionDto
          description: positionDto
          required: true
          schema:
            $ref: '#/definitions/PositionDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PositionDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /position/update:
    put:
      tags:
        - positionController
      summary: update position
      operationId: updatePosition
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: positionDto
          description: positionDto
          required: true
          schema:
            $ref: '#/definitions/PositionDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PositionDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/position/{positionId}':
    get:
      tags:
        - positionController
      summary: get position by id
      operationId: getPosition
      produces:
        - '*/*'
      parameters:
        - name: positionId
          in: path
          description: positionId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PositionDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /sick/new:
    post:
      tags:
        - sickLeaveController
      summary: insert new sick leave
      operationId: insertSickLeave
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: sickLeaveDto
          description: sickLeaveDto
          required: true
          schema:
            $ref: '#/definitions/SickLeaveDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SickLeaveDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /sick/update:
    put:
      tags:
        - sickLeaveController
      summary: update sick leave
      operationId: updateSickLeave
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: sickLeaveDto
          description: sickLeaveDto
          required: true
          schema:
            $ref: '#/definitions/SickLeaveDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SickLeaveDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/sick/{sickId}':
    get:
      tags:
        - sickLeaveController
      summary: get sick leave by id
      operationId: getSickLeave
      produces:
        - '*/*'
      parameters:
        - name: sickId
          in: path
          description: sickId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SickLeaveDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /vacation/employee_last:
    get:
      tags:
        - vacationController
      summary: get last vacation by employee
      operationId: getLastVacationByEmployee
      produces:
        - '*/*'
      parameters:
        - name: employeeId
          in: query
          description: employeeId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/VacationDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /vacation/new:
    post:
      tags:
        - vacationController
      summary: insert new Vacation
      operationId: insertVacation
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: vacationDto
          description: vacationDto
          required: true
          schema:
            $ref: '#/definitions/VacationDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/VacationDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /vacation/update:
    put:
      tags:
        - vacationController
      summary: update vacation
      operationId: updateVacation
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: vacationDto
          description: vacationDto
          required: true
          schema:
            $ref: '#/definitions/VacationDto'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/VacationDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/vacation/{vacationId}':
    get:
      tags:
        - vacationController
      summary: get vacation by id
      operationId: getVacation
      produces:
        - '*/*'
      parameters:
        - name: vacationId
          in: path
          description: vacationId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/VacationDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
...
  PositionDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    title: PositionDto
  SickLeaveDto:
    type: object
    properties:
      dateEnd:
        type: string
        format: date
      dateStart:
        type: string
        format: date
      employeeId:
        type: integer
        format: int64
      id:
        type: integer
        format: int64
      typeOfSickLeave:
        type: string
        enum:
          - BASIC
          - FAMILY_MEMBER
          - CHILD
    title: SickLeaveDto
  VacationDto:
    type: object
    properties:
      dateEnd:
        type: string
        format: date
      dateStart:
        type: string
        format: date
      employeeId:
        type: integer
        format: int64
      id:
        type: integer
        format: int64
      typeOfVacation:
        type: string
        enum:
          - ANNUAL_VACATION
          - ADDITIONAL_VACATION
    title: VacationDto